requirements.txt
# ChromaDB Dependencies
chromadb>=0.4.0

# Legacy Azure Dependencies (keeping for reference)
azure-search-documents==11.4.0
azure-core==1.29.5

# Core Dependencies
openai>=1.0.0,<2.0.0
python-frontmatter==1.1.0
python-dotenv==1.0.0
python-docx==1.2.0
mcp>=1.0.0
